Fix Copy Paste  (#1222)
Address an issue where copying from a datagrid in FDA would not paste
into Excel. I have not been able to track down when, or why this
happened. "Copy" was handled by the System.Windows Library, not us. My
best guess is that the behavior changed between Net 6.0 -< Net 9.0, but
I have not yet found the commit where we changed, or found discussion on
the webs.

Nevertheless, this is a pretty clean fix as far as hacks go. Windows
copies data in multiple formats when you call copy. I grab the data in
"text" format, purge the clipboard, and set the text data back to the
clipboard every copy, so the data we want for excel is the only data in
there.

seems to work.