Geospatial update (#1221)
This update moves us largely off RasMapper for geospatial operations. We
do still use PointM as our coordinate for structures on construction.
That's important because we need to work in old ras space (RasMapper) to
use RAS6.x terrains and results. So everywhere we can, we're now
sticking to the actively developed RAS2025 codebase and ways of doing
things, and only reverting for operations we can't support.