Fix uncertainty around graphical stage frequency  (#1223)
I have redesigned the computational algorithm for generating uncertainty
about graphical frequency curves and have added a unit test for the edge
case (provided by the Algiers PDT) that illuminated the need for the
change. HEC-FDA Version 2.0 is now producing nearly exactly the same
uncertainty as HEC-FDA Version 1.4.3 for this edge case.

---------

Co-authored-by: Brennan Beam <64556723+Brennan1994@users.noreply.github.com>