Debug tools (#1292)
This PR adds some nice debugging tools, including ScottPlots and an
ASCII histogram visualizer.

It also Rips out the less simple method into a static method in a
separate class. Doing this for ease of review. I also removed some
parameters that used to be configurable from it, because I feel
philosophically they shouldn't be. That method is to be as rigidly the
same as 1.4 as possible.

Critically, computation responsibility moved form GraphicalDistribution
-> GraphicalFrequencyCalculators. Those two classes should be the beef
of this review.